+++ lib/slack/web/client.rb
@@ -12,6 +12,16 @@ module Slack
           send("#{key}=", options[key] || Slack::Web.config.send(key))
         end
       end
+
+      class << self
+        def configure
+          block_given? ? yield(Config) : Config
+        end
+
+        def config
+          Config
+        end
+      end
     end
   end
 end