+++ lib/slack/web/config.rb
@@ -15,7 +15,7 @@ module Slack
 
       def reset
         self.endpoint = 'https://slack.com/api/'
-        self.user_agent = "Slack Ruby Client #{Slack::VERSION}"
+        self.user_agent = "Slack Ruby Client/#{Slack::VERSION}"
         self.ca_path = `openssl version -a | grep OPENSSLDIR | awk '{print $2}'|sed -e 's/\"//g'`
         self.ca_file = "#{ca_path}/ca-certificates.crt"
       end